requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.22
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
pytorch==2.1.0
transformers==4.34.1
python-dotenv==1.0.0
geopy==2.4.0
ortools==9.7.2996
psycopg2-binary==2.9.9
jinja2==3.1.2
python-docx==1.0.0
reportlab==4.0.6